♻️ Refactor(infra): src/db → src/infra/db 레이어 분리
DB 인프라를 infra/db/ 레이어로 격리.
app/, components/, proxy.ts의 @/db/ import를 @/infra/db/ 로 일괄 교체.
drizzle.config.ts 스키마 경로도 업데이트.
