+++ src/app/categories/page.tsx
@@ -1,5 +1,5 @@
-import { getRepositories } from "@/db/repositories";
-import type { CategoryData } from "@/db/types";
+import { getRepositories } from "@/infra/db/repositories";
+import type { CategoryData } from "@/infra/db/types";
 import { CategoryList } from "@/components/CategoryList";
 import { Metadata } from "next";
