+++ src/infra/db/repositories/CategoryRepository.ts
