+++ src/infra/db/repositories/SyncLogRepository.ts
