+++ src/infra/db/schema/categories.ts
